coord: absorb catalog crate
The dependency graph now looks like:

    SQL <- coord

This is the bare minimum code movement necessary to make this viable.
Many refactoring commits will follow.
