--- src/catalog/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "catalog"
-description = "Metadata storage."
-version = "0.1.0"
-edition = "2018"
-publish = false
-
-[lib]
-path = "lib.rs"
-
-[dependencies]
-backtrace = "0.3.43"
-bincode = { version = "1.2", optional = true }
-chrono = "0.4"
-dataflow-types = { path = "../dataflow-types" }
-expr = { path = "../expr" }
-failure = "0.1.8"
-lazy_static = "1.4.0"
-log = "0.4.8"
-itertools = "0.9"
-ore = { path = "../ore" }
-rand = "0.7.3"
-regex = "1"
-repr = { path = "../repr" }
-rusqlite = { version = "0.23", features = ["bundled"] }
-serde = "1.0"
-serde_json = "1.0.52"
