+++ src/coord/Cargo.toml
@@ -11,9 +11,10 @@ path = "lib.rs"
 [dependencies]
 avro = { path = "../avro" }
 aws-util = { path = "../aws-util" }
-catalog = { path = "../catalog" }
-chrono = "0.4"
+backtrace = "0.3.43"
+bincode = { version = "1.2", optional = true }
 ccsr = { path = "../ccsr" }
+chrono = "0.4"
 comm = { path = "../comm" }
 dataflow = { path = "../dataflow" }
 dataflow-types = { path = "../dataflow-types" }
@@ -29,17 +30,18 @@ log = "0.4"
 ore = { path = "../ore" }
 pgrepr = { path = "../pgrepr" }
 prometheus = { git = "https://github.com/MaterializeInc/rust-prometheus.git", default-features = false }
-repr = { path = "../repr" }
+rand = "0.7.3"
 rdkafka = { git = "https://github.com/fede1024/rust-rdkafka.git", features = ["cmake-build", "libz-static"] }
 regex = "1.3.4"
+repr = { path = "../repr" }
 rusoto_kinesis = "0.43.0"
 rusqlite = { version = "0.23", features = ["bundled", "unlock_notify"] }
-serde = "1"
+serde = "1.0"
 serde_json = "1.0.52"
 sql = { path = "../sql" }
 sql-parser = { path = "../sql-parser" }
 symbiosis = { path = "../symbiosis" }
 timely = { git = "https://github.com/TimelyDataflow/timely-dataflow", features = ["bincode"] }
-transform = { path = "../transform" }
 tokio = "0.2"
+transform = { path = "../transform" }
 url = "2"
