+++ src/coord/coord.rs
@@ -31,8 +31,6 @@ use futures::stream::{self, StreamExt, TryStreamExt};
 use timely::progress::frontier::Antichain;
 use timely::progress::ChangeBatch;
 
-use catalog::names::{DatabaseSpecifier, FullName};
-use catalog::{Catalog, CatalogItem, PlanContext, SinkConnectorState};
 use dataflow::logging::materialized::MaterializedEvent;
 use dataflow::{SequencedCommand, WorkerFeedback, WorkerFeedbackWithMeta};
 use dataflow_types::logging::LoggingConfig;
@@ -48,11 +46,13 @@ use ore::collections::CollectionExt;
 use ore::thread::JoinHandleExt;
 use repr::{ColumnName, Datum, RelationDesc, RelationType, Row};
 use sql::{
-    ExplainOptions, MutationKind, ObjectType, Params, Plan, PreparedStatement, Session, Statement,
+    DatabaseSpecifier, ExplainOptions, FullName, MutationKind, ObjectType, Params, Plan,
+    PlanContext, PreparedStatement, Session, Statement,
 };
 use sql_parser::ast::ExplainStage;
 use transform::Optimizer;
 
+use crate::catalog::{self, Catalog, CatalogItem, SinkConnectorState};
 use crate::persistence::SqlSerializer;
 use crate::timestamp::{TimestampConfig, TimestampMessage, Timestamper};
 use crate::util::ClientTransmitter;
