+++ src/coord/lib.rs
@@ -20,6 +20,7 @@
 //! [`pgwire`](../pgwire/index.html) produces, though they can, in theory, be
 //! provided by something other than a pgwire server.
 
+mod catalog;
 mod command;
 mod coord;
 mod persistence;
