+++ src/coord/timestamp.rs
@@ -33,7 +33,6 @@ use rusoto_kinesis::KinesisClient;
 use rusqlite::{params, NO_PARAMS};
 
 use aws_util;
-use catalog::sql::SqlVal;
 use dataflow::source::read_file_task;
 use dataflow::source::FileReadStyle;
 use dataflow_types::{
@@ -43,6 +42,7 @@ use dataflow_types::{
 use expr::{PartitionId, SourceInstanceId};
 use ore::collections::CollectionExt;
 
+use crate::catalog::sql::SqlVal;
 use crate::coord;
 
 /// Number of seconds that must elapse before we update watermarks again
@@ -547,7 +547,7 @@ pub struct Timestamper {
     byo_sources: HashMap<SourceInstanceId, ByoTimestampConsumer>,
 
     // Connection to the underlying SQL lite instance
-    storage: Arc<Mutex<catalog::sql::Connection>>,
+    storage: Arc<Mutex<crate::catalog::sql::Connection>>,
 
     tx: futures::channel::mpsc::UnboundedSender<coord::Message>,
     rx: std::sync::mpsc::Receiver<TimestampMessage>,
@@ -742,7 +742,7 @@ fn identify_consistency_format(enc: DataEncoding, env: Envelope) -> ConsistencyF
 impl Timestamper {
     pub fn new(
         config: &TimestampConfig,
-        storage: Arc<Mutex<catalog::sql::Connection>>,
+        storage: Arc<Mutex<crate::catalog::sql::Connection>>,
         tx: futures::channel::mpsc::UnboundedSender<coord::Message>,
         rx: std::sync::mpsc::Receiver<TimestampMessage>,
     ) -> Self {
@@ -785,7 +785,7 @@ impl Timestamper {
         }
     }
 
-    fn storage(&self) -> MutexGuard<catalog::sql::Connection> {
+    fn storage(&self) -> MutexGuard<crate::catalog::sql::Connection> {
         self.storage.lock().expect("lock poisoned")
     }
