+++ src/sql/Cargo.toml
@@ -11,8 +11,6 @@ path = "lib.rs"
 [dependencies]
 avro = { path = "../avro" }
 aws-arn = "0.1.1"
-# TODO(benesch): eventually only the tests will need the bincode feature.
-catalog = { path = "../catalog", features = ["bincode"] }
 ccsr = { path = "../ccsr" }
 chrono = "0.4"
 dataflow-types = { path = "../dataflow-types" }
@@ -30,6 +28,7 @@ regex = "1.3.7"
 repr = { path = "../repr" }
 reqwest = { version = "0.10.4" }
 rusoto_core = "0.43.0-beta.1"
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sql-parser = { path = "../sql-parser" }
 tokio = { version = "0.2.20", features = ["fs"] }
