+++ src/sql/tests/parameters.rs
@@ -9,14 +9,14 @@
 
 use std::error::Error;
 
-use catalog::{BincodeSerializer, Catalog};
 use ore::collections::CollectionExt;
 use pgrepr::Type;
+use sql::catalog::DummyCatalog;
 use sql::Session;
 
 #[test]
 fn test_parameter_type_inference() -> Result<(), Box<dyn Error>> {
-    let catalog = Catalog::open::<BincodeSerializer, _>(None, |_| ())?;
+    let catalog = DummyCatalog;
     let session = Session::default();
     let test_cases = vec![
         (
