+++ src/symbiosis/Cargo.toml
@@ -9,7 +9,6 @@ publish = false
 path = "lib.rs"
 
 [dependencies]
-catalog = { path = "../catalog" }
 chrono = { version = "0.4", features = ["serde"] }
 dataflow-types = { path = "../dataflow-types" }
 expr = { path = "../expr" }
